RDMA/bnxt_re: Move the UAPI methods to a dedicated file
[ Upstream commit eee6268421a2ccc6d47d8e175a1f4bfcd78a83ca ]

This is in preparation for upcoming patches in the series.
Driver has to support additional UAPIs for some applications.
Moving current UAPI implementation to a new file, uapi.c.

Link: https://patch.msgid.link/r/20260302110036.36387-2-sriharsha.basavapatna@broadcom.com
Stable-dep-of: a57592c6392a ("RDMA/bnxt_re: Fail DBR related page allocation UAPIs if the feature is disabled")
(cherry picked from commit 0fe155aa844e44157b82b7390f15b2a49a59765c)
